update javascript example
