+++ JavaScript/testing-sdk/src/run.ts
@@ -1,12 +1,11 @@
 import * as flashcardGenerator from './test-suites/flashcard-generator';
 import * as studyGuideOutline from './test-suites/study-guide-outline';
+import * as flashcardGeneratorManaged from './test-suites/flashcard-generator-managed';
 
 (async () => {
   await Promise.all([
     flashcardGenerator.run(),
     studyGuideOutline.run(),
-    // Uncomment after setting up managed test cases
-    // for flashcard-generator
-    // flashcardGenerator.runWithManagedTestCases(),
+    flashcardGenerator.run(),
   ]);
 })();
